+++ oysterpack-lib-template/src/tests.rs
@@ -6,22 +6,58 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
+//! unit test support
+
+extern crate fern;
 use build;
+use chrono;
+use log;
+use std::io;
+
+fn init_logging() -> Result<(), fern::InitError> {
+    fern::Dispatch::new()
+        .format(|out, message, record| {
+            out.finish(format_args!(
+                "{}[{}][{}] {}",
+                chrono::Local::now().format("[%Y-%m-%d][%H:%M:%S%.6f]"),
+                record.level(),
+                record.target(),
+                message
+            ))
+        }).level(log::LevelFilter::Warn)
+        // TODO: update module name
+        .level_for("oysterpack_lib_template", log::LevelFilter::Debug)
+        .chain(io::stdout())
+        .apply()?;
+
+    Ok(())
+}
+
+lazy_static! {
+    pub static ref INIT_FERN: Result<(), fern::InitError> = init_logging();
+}
+
+pub fn run_test<F: FnOnce() -> ()>(test: F) {
+    let _ = *INIT_FERN;
+    test()
+}
 
 #[test]
 fn build_info() {
-    println!("{}", concat!(env!("OUT_DIR"), "/built.rs"));
-    println!(
-        "This is version {}{}, built for {} by {}.",
-        build::PKG_VERSION,
-        build::GIT_VERSION.map_or_else(|| "".to_owned(), |v| format!(" (git {})", v)),
-        build::TARGET,
-        build::RUSTC_VERSION
-    );
-    println!(
-        "I was built with profile \"{}\", features \"{}\" on {}",
-        build::PROFILE,
-        build::FEATURES_STR,
-        build::BUILT_TIME_UTC
-    );
+    run_test(|| {
+        info!("{}", concat!(env!("OUT_DIR"), "/built.rs"));
+        info!(
+            "This is version {}{}, built for {} by {}.",
+            build::PKG_VERSION,
+            build::GIT_VERSION.map_or_else(|| "".to_owned(), |v| format!(" (git {})", v)),
+            build::TARGET,
+            build::RUSTC_VERSION
+        );
+        info!(
+            "I was built with profile \"{}\", features \"{}\" on {}",
+            build::PROFILE,
+            build::FEATURES_STR,
+            build::BUILT_TIME_UTC
+        );
+    });
 }