+++ oysterpack-uid/src/lib.rs
@@ -19,14 +19,15 @@
 #![deny(missing_docs, missing_debug_implementations)]
 #![doc(html_root_url = "https://docs.rs/oysterpack_uid/0.1.0")]
 
-extern crate built;
+#[macro_use]
+extern crate oysterpack_built_mod;
 #[macro_use]
 extern crate log;
 #[macro_use]
 extern crate lazy_static;
 extern crate chrono;
 
-pub mod build;
+op_build_mod!();
 
 #[cfg(test)]
 mod tests;