+++ oysterpack-built/CHANGELOG.md
@@ -6,6 +6,14 @@ This project adheres to [Semantic Versioning](http://semver.org/), as described 
 
 ## Unreleased
 
+## \[0.2.1\] 2018-09-23
+
+### Added
+- copyright to tests/version-numbers.rs
+
+### Removed
+- dev-dependencies on semver and chrono
+
 ## \[0.2.0\] 2018-09-23
 
 ### Changed